% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdtomonthly.R
\name{acdtomonthly}
\alias{acdtomonthly}
\title{Function to turn UCDP annual conflict data into a conflict-month dataset}
\usage{
acdtomonthly(
  postwar_months = 0,
  include_alqaida = FALSE,
  start_year = NULL,
  end_year = NULL
)
}
\arguments{
\item{postwar_months}{Maximum postwar months to include, default is only active months. Put 5000 if you want all possible postwar months in the data.}

\item{include_alqaida}{Optional: Include conflict 418 (US-al-Qaida), default is to exclude.}

\item{start_year}{Optional: Specify the start year. Default: 1989.}

\item{end_year}{Specify the end year. Default: longest possible in version (currently 2023).}
}
\value{
Basedata, a conflict-month dataset
}
\description{
Function to turn UCDP annual conflict data into a conflict-month dataset
}
\examples{
acdtomonthly(postwar_months = 5000)
acdtomonthly(start_year = 2015, end_year = 2023,include_alqaida = TRUE,postwar_months = 24)
acdtomonthly()
}
