% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actorlink.R
\name{actorlink}
\alias{actorlink}
\title{Links those forms of violence to the conflict-month dataset that can be linked by actor identifiers}
\usage{
actorlink(gedprepped, basedata)
}
\arguments{
\item{gedprepped}{The GED event dataset, prepared in function "gedprep"}

\item{basedata}{The conflict-month dataset, prepared in function "acdtomonthly"}
}
\value{
The basedata with violence added, and a version of GED that tracks for each event how it was linked
}
\description{
Links those forms of violence to the conflict-month dataset that can be linked by actor identifiers
}
\examples{
\dontrun{
  actorlink(gedprepped,basedata)
}
}
\keyword{internal}
