% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{finalize}
\alias{finalize}
\title{Cleans up final dataset}
\usage{
finalize(basedata)
}
\arguments{
\item{basedata}{}
}
\value{
The final product of our R package: conflict-month dataset with all forms of violence linked
}
\description{
Cleans up final dataset
}
\examples{
\dontrun{
  finalize(basedata)
}
}
\keyword{internal}
