% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.R
\name{intersect}
\alias{intersect}
\title{Intersects the GED event dataset with the conflict zones for later spatial linking}
\usage{
intersect(gedprepped, episode_zones)
}
\arguments{
\item{gedprepped}{Data passed on from gedprep}

\item{episode_zones}{Data passed on from makezones}
}
\value{
A reduced GED event dataset where each event has one or several conflict zones assigned (one event = several rows if it falls in more than one zone)
}
\description{
Intersects the GED event dataset with the conflict zones for later spatial linking
}
\examples{
\dontrun{
  intersect(gedprepped,episode_zones)
}
}
\keyword{internal}
