% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makezones.R
\name{makezones}
\alias{makezones}
\title{Creates a conflict zone (spatial object) for each episode present in the user-created conflict-month dataset}
\usage{
makezones(basedata, buffer_percent = 0, clipcountry = TRUE)
}
\arguments{
\item{basedata}{Passes the conflict-month dataset on}

\item{buffer_percent}{How much buffer to add around each conflict zone (in percent, numeric from 0 to 100). Default is no buffer.}

\item{clipcountry}{Whether to clip conflict zones at the borders of the country/countries primarily affected by the conflict. Default is yes.}
}
\value{
The spatial objects, and the conflict-month data with a column added on whether there is a zone available for that episode
}
\description{
Creates a conflict zone (spatial object) for each episode present in the user-created conflict-month dataset
}
\examples{
\dontrun{
  makezones(basedata, clipcountry = FALSE, buffer_percent = 20)
}
\dontrun{
  makezones(basedata)
}
\dontrun{
  makezones(basedata, clipcountry = TRUE)
}
}
\keyword{internal}
