% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiallink.R
\name{spatiallink}
\alias{spatiallink}
\title{Links those forms of violence to the conflict-month dataset that have to be linked by location}
\usage{
spatiallink(gedintersection, basedata, gedtrack, divide_deaths = FALSE)
}
\arguments{
\item{gedintersection}{GED events spatially intersected with conflict zones. Passed on from the intersect function}

\item{basedata}{Conflict-month dataset, generated by the acdtomonthly function and passed through}

\item{gedtrack}{A version of GED that tracks for each event how it was linked}

\item{divide_deaths}{User choice for events that fall into multiple conflict zones: use full casualties in each or divide deaths between the conflicts?}
}
\value{
The basedata with violence added, and a version of GED that tracks for each event how it was linked
}
\description{
Links those forms of violence to the conflict-month dataset that have to be linked by location
}
\examples{
\dontrun{
  spatiallink(gedintersection,basedata,gedtrack,divide_deaths = FALSE)
}
}
\keyword{internal}
