% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplitcb.R
\name{strsplitcb}
\alias{strsplitcb}
\title{Split a string}
\usage{
strsplitcb(x, split)
}
\arguments{
\item{x}{A character vector with one element.}

\item{split}{What to split on.}
}
\value{
A character vector.
}
\description{
Split a string
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplitcb(x, split = ",")
}
