% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdplink.R
\name{ucdplink}
\alias{ucdplink}
\title{Creates a UCDP conflict-month dataset with information on various forms of violence from UCDP GED, linked via actors or location.}
\usage{
ucdplink(
  postwar_months = 0,
  include_alqaida = FALSE,
  start_year = NULL,
  end_year = NULL,
  buffer_percent = 0,
  clipcountry = TRUE,
  divide_deaths = FALSE,
  include_gedtrack = FALSE
)
}
\arguments{
\item{postwar_months}{How many inactive months to add to each episode. Default is 0 (only active conflict). Add 5000 if you want all possible postwar months until the end of the dataset.}

\item{include_alqaida}{Whether to include conflict 418 (US-al-Qaida), default is to exclude, as this conflict did not take place on US territory.}

\item{start_year}{Specify the start year. Default: 1989.}

\item{end_year}{Specify the end year. Default: longest possible (currently 2023).}

\item{buffer_percent}{How much buffer to add around each conflict zone (in percent, numeric from 0 to 100). Default is no buffer.}

\item{clipcountry}{Whether to clip conflict zones at the borders of the country/countries primarily affected by the conflict. Default is to clip.}

\item{divide_deaths}{For events that fall into multiple conflict zones: use full casualties in each or divide deaths between the conflicts? Default is to use full.}

\item{include_gedtrack}{Whether to also return a version of GED in which event has information on how it was linked to the conflict-month dataset. Default is FALSE.}
}
\value{
A conflict-month dataset with casualty counts from various forms of violence
}
\description{
Creates a UCDP conflict-month dataset with information on various forms of violence from UCDP GED, linked via actors or location.
}
\note{
The code in this package is licensed under GPL-3, but the final dataset produced is licensed under CC BY 4.0.

When using this function, please ensure that you correctly attribute the source data:

UCDP/PRIO Armed Conflict Dataset version 24.1 from the Uppsala Conflict Data Program.

UCDP Georeferenced Event Dataset (GED) Global version 24.1 from the Uppsala Conflict Data Program.

For UCDP citation details, refer to the UCDP website: https://ucdp.uu.se/downloads/

WZONE: Zones of Armed Conflicts v9.0

For WZONE citation details see https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/PUWJEU
}
\examples{
ucdplink(postwar_months = 0, include_alqaida = FALSE, start_year = 1999, end_year = 2010,
buffer_percent = 0, clipcountry = TRUE, divide_deaths = FALSE, include_gedtrack = FALSE)
}
