% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedprep.R
\name{gedprep}
\alias{gedprep}
\title{Takes UCDP GED and processes it to:
a) divide multi-month events into separate one-month events (fatalities evenly divided between them)
b) divide events by coalition actors to the coalition-constituting actors (fatalities evenly divided between them)}
\usage{
gedprep()
}
\value{
UCDP GED processed (with more rows than the original)
}
\description{
Takes UCDP GED and processes it to:
a) divide multi-month events into separate one-month events (fatalities evenly divided between them)
b) divide events by coalition actors to the coalition-constituting actors (fatalities evenly divided between them)
}
\examples{
gedprep()
}
