% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdplink.R
\name{ucdplink}
\alias{ucdplink}
\title{Creates a UCDP conflict-month dataset with information on various forms of violence from UCDP GED, linked via actors or space}
\usage{
ucdplink(
  postwar_months = 0,
  include_alqaida = FALSE,
  start_year = NULL,
  end_year = NULL,
  buffer_percent = 0,
  clipcountry = TRUE,
  divide_deaths = FALSE,
  include_gedtrack = FALSE
)
}
\arguments{
\item{postwar_months}{How many inactive months to add to each episode. Default is 0 (only active conflict). Add 5000 if you want all possible postwar months until the end of the dataset.}

\item{include_alqaida}{Whether to include conflict 418 (US-al-Qaida), default is to exclude, as this conflict did not take place on US territory.}

\item{start_year}{Specify the start year. Default: 1989.}

\item{end_year}{Specify the end year. Default: longest possible (currently 2023).}

\item{buffer_percent}{How much buffer to add around each conflict zone (in percent, numeric from 0 to 100). Default is no buffer.}

\item{clipcountry}{Whether to clip conflict zones at the borders of the country/countries primarily affected by the conflict. Default is to clip.}

\item{divide_deaths}{For events that fall into multiple conflict zones: use full casualties in each or divide deaths between the conflicts? Default is to use full.}

\item{include_gedtrack}{Whether to also return a version of GED in which event has information on how it was linked to the conflict-month dataset.}
}
\value{
A conflict-month dataset based on UCDP conflict data, with casualty counts from various forms of violence from UCDP GED
}
\description{
Creates a UCDP conflict-month dataset with information on various forms of violence from UCDP GED, linked via actors or space
}
\examples{
ucdplink(include_alqaida = FALSE, start_year = 1999, end_year = 2010, include_gedtrack = TRUE)
\dontrun{
  ucdplink(postwar_months = 24, buffer_percent = 20, clipcountry = FALSE, divide_deaths = TRUE)
}
}
